# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Boat reservation application API.",
        "title": "Harbour Wave API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/availability/{month}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns daily availability for a given month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availability"
                ],
                "summary": "Get monthly availability",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Month in YYYY-MM format",
                        "name": "month",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AvailabilityMonthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        },
        "/bookings": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a pending booking and holds slot capacity. Requires idempotency key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Create a new booking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Idempotency Key",
                        "name": "X-Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Booking payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.CreateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.CreateBookingResponse"
                        }
                    },
                    "400": {
                        "description": "INVALID_INPUT",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "NOT_AUTHENTICATED",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "409": {
                        "description": "SLOT_TAKEN",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "422": {
                        "description": "VALIDATION_FAILED, DATE_BLOCKED, or SLOT_BLOCKED",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "BOOKINGS_DISABLED or SERVICE_UNAVAILABLE",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        },
        "/bookings/by-session/{sessionId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns minimal status info while pending, or the full booking once confirmed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Get booking by session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.BookingBySessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        },
...
        "/payment/webhook": {
            "post": {
                "description": "Receives payment status updates via HMAC-secured webhook.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Payment gateway webhook",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "INVALID_SIGNATURE or CANNOT_READ_BODY",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "WEBHOOK_PROCESSING_FAILED",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        },
        "/slots/{date}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all available and blocked slots for a specific date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availability"
                ],
                "summary": "Get slots for a date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.SlotsForDateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_harbour-wave_harbour-wave-backend_internal_model.AvailabilityDay": {
            "type": "object",
            "properties": {
                "availableSlots": {
                    "type": "integer"
                },
                "blocked": {
                    "type": "boolean"
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.AvailabilityMonthResponse": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.AvailabilityDay"
                    }
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.BookingBySessionResponse": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.BookingPublicView"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.BookingPublicView": {
            "type": "object",
            "properties": {
                "contact": {
                    "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.ContactPublicView"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "quantities": {
                    "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.Quantities"
                },
                "status": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "number"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.ContactPublicView": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.ContactRequest": {
            "type": "object",
            "required": [
                "firstName",
                "lastName",
                "phone"
            ],
            "properties": {
                "email": {
                    "description": "ignored; JWT email used",
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "lastName": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "phone": {
                    "description": "REQUIRED",
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.CreateBookingRequest": {
            "type": "object",
            "required": [
                "contact",
                "date",
                "quantities",
                "time"
            ],
            "properties": {
                "contact": {
                    "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.ContactRequest"
                },
                "date": {
                    "type": "string"
                },
                "quantities": {
                    "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.QuantitiesRequest"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.CreateBookingResponse": {
            "type": "object",
            "properties": {
                "bookingId": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "number"
                }
            }
        },
...
        "github_com_harbour-wave_harbour-wave-backend_internal_model.Quantities": {
            "type": "object",
            "properties": {
                "big": {
                    "type": "integer"
                },
                "child": {
                    "type": "integer"
                },
                "medium": {
                    "type": "integer"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.QuantitiesRequest": {
            "type": "object",
            "properties": {
                "big": {
                    "type": "integer"
                },
                "child": {
                    "type": "integer"
                },
                "medium": {
                    "type": "integer"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.SlotForDate": {
            "type": "object",
            "properties": {
                "availableBig": {
                    "type": "integer"
                },
                "availableMedium": {
                    "type": "integer"
                },
                "blocked": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                },
                "totalBig": {
                    "type": "integer"
                },
                "totalMedium": {
                    "type": "integer"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_internal_model.SlotsForDateResponse": {
            "type": "object",
            "properties": {
                "bookingsEnabled": {
                    "type": "boolean"
                },
                "date": {
                    "type": "string"
                },
                "dateBlocked": {
                    "type": "boolean"
                },
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_internal_model.SlotForDate"
                    }
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_pkg_httpx.ErrorBody": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "issues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_harbour-wave_harbour-wave-backend_pkg_httpx.Issue"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_harbour-wave_harbour-wave-backend_pkg_httpx.Issue": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}